Added newer files and made the ptrs uint8_t

+++ src/gpio.h
@@ -9,15 +9,15 @@ class gpio {
 protected:
   uint8_t flags;
   const uint8_t pin;
-  const uint16_t readBank;
+  const uint8_t readBank;
   const uint8_t readOffset;
-  const uint16_t writeBank;
+  const uint8_t writeBank;
   const uint8_t writeOffset;
-  constexpr uint16_t getReadBank(uint8_t pin);
+  constexpr uint8_t getReadBank(uint8_t pin);
   constexpr uint8_t getReadOffset(uint8_t pin);
-  constexpr uint16_t getWriteBank(uint8_t pin);
+  constexpr uint8_t getWriteBank(uint8_t pin);
   constexpr uint8_t getWriteOffset(uint8_t pin);
-  constexpr uint16_t getDDRBank(uint8_t pin);
+  constexpr uint8_t getDDRBank(uint8_t pin);
   constexpr uint8_t getDDROffset(uint8_t pin);
 
 public:
